requirements.txt
# Core libraries
numpy>=1.26.4
pandas>=2.2.2

# Visualization
matplotlib>=3.8.4
seaborn>=0.12.2
wordcloud>=1.9.3

# NLP tools
nltk>=3.8.1
textblob>=0.18.0.post0
vaderSentiment>=3.3.2
emoji>=2.10.0
contractions>=0.1.73
spacy>=3.7.2
symspellpy>=6.7.7
beautifulsoup4>=4.12.3

# HuggingFace ecosystem
torch>=2.0
transformers>=4.40.0
datasets>=2.18.0
evaluate>=0.4.1

# Model Training & Evaluation
scikit-learn>=1.4.2
imbalanced-learn>=0.12.0

# Gradient Boosting
xgboost>=2.0.3

# YAML
pyyaml>=6.0.1

# Web App
streamlit>=1.34.0

# Progress bar (optional but useful)
tqdm>=4.66.4

# Web scrapping
requests